added well-known chatty, submissive, or empty words and phrases
I've added many conversational, chatty, submissive/passive, empty/vacuous words and phrases to the list which often undermine and weaken the message one is attempting to convey. They're all common phrases and words that I've self-censored myself (as a male), grammar-reviewed and suggested to other men as well as women when I've been asked to proofread high-stakes emails or replies.

On the other hand, there's also a fine line here to tread: sometimes mimicking a superior's conversational or chatty choice of language when the stakes are low is, in my opinion, totally acceptable in moderation. In fact, mimicking another person's (body language) and word/language choice is a commonly understood strategy to establish rapport.

So - in summary, context matters. Not all "sorry-not-sorry" type language needs to be self-censored entirely, but otherwise this tool is a fantastic idea and I hope to see the collection of warnings grow and extend to other programs and interfaces, like desktop email programs (Microsoft Outlook) and office collaboration chat tools such as Slack.

Cheers!