fix: keyword '!!!' was not working with plugin
Wasn't working because regex was not recording a match with \b and \b as word boundaries for ! which
really denotes the end of a sentence. Thought process behind this fix was to separate phrases from
end of sentence punctuations. I added a quantifier in the json file to cover 3 or more exclamation
points. Not sure if this is the way forward but it gets us a fix for the !!! keyword