fix: fix leaky divs
previously when user would click out of message box and back in if you check the html dom in the
inspector you will see an extra div getting appended each time. This is because setting state is
unreliable when not using a function parameter vs setState with an object. see
https://reactjs.org/docs/faq-state.html