fix: correctly trigger warnings when matched text is on separate lines
Match was occurring on separate lines because we were searching across the entire container div
rather than the individual divs that each line in the mssage is made up of. This fix iterates
through each childNode (div) of text and identifies a match that way. Note, this doesn't fix the
highlighting though of phrases that span multiple lines.