refactor: a refactor / performance change that makes the JustNotSorry component more 'class-like'
This is a big refactor that was part experimental and has turned into a bigger commit than I
would've liked. However it seems to bring us back to a more similar performance-ability as in
v1.6.3. Was able to remove some extra observers and thin out the class overall. Also changes such as
first collating a list of warnings and setting the state once rather than multiple times when a
method is called. This however involved pulling in the dom-regexp-match function and modifying it
slightly.