fix(performance): changes to improve performance
I believe there are two aspects that affected the performance of JNS since v1.6.3 - one was the
mutations. This has been 'improved' by introducing a set object to make sure event listeners are not
added to existing nodes. Another improvement was to refactor out the parentNode call before building
the array of WarningHighlights.