fix(outlook.com): fix positioning of underlines on outlook.com
This fixes an issue with the underlines in emails on outlook.com appearing above the words.
The root cause was that outlook.com changed the DOM structure, adding padding to the contenteditable div that was the normal parent
as well as adding another containing div with no `position` attribute to the hierarchy.
The new intermediate was being used as the parent for calculating coords but the coords of the words were relative to the contenteditable div.
Using offsetParent instead of parentNode skips the intermediate div and makes things line up properly.

fixes #134