feat(privacy): remove google analytics tracking
Remove JNS's connection to Google Analytics. This sets the stage for an
easier migration to v3 of the chrome extension manifest and allows us to
narrow the permissions requested by JNS.

The previously used Google Analytics tracking using
https://github.com/googlearchive/chrome-platform-analytics only reported a custom event when the user
installed JustNotSorry. The Chrome Web Store also gathers this information, so we're not losing anything.